Replaced SUBA #$FF with BITA #$07 in joystick button check.
The code for checking for joystick buttons was causing intermittent failures.  Occasionally the result of the SUBA would not be 0 even when joystick buttons were not pressed.  The change made here is what I should have done in the first place.  It masks off any keyboard input by writing $FF to the PIA ($FF02).  Joystick button presses will clear bits in $FF00 even when keyboard input is blocked.  Joystick button bits are 0-3 in $FF00.  This code now checks if ANY of those bits are clear with the BITA #$07.  If they are clear then keyboard input is not allowed.